requirements.txt
asgiref==3.8.1
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.25
django-allauth==0.50.0
django-cors-headers==4.3.1
django-filter==23.5
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
oauthlib==3.2.2
psycopg2==2.9.9
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.2
requests-oauthlib==2.0.0
sqlparse==0.5.1
